Added the Fast Inverse Square Root method
This method calculates the inverse square root of a number in one third of the time as the built in C++ functions. It only has around a one percent error and uses faster functions such as add and multiply rather that sqrt() and divide. It is very useful in normalizing vectors for video games.